Rework the way we look at the end boundary when looking an element in a parent
The test `MaxDataSize >= (PossibleID_Length + PossibleSizeLength + SizeFound)`
is incorrect when there was garbage data skipped inside the PossibleIdNLength
table.

Now we keep track of how many memmove we had to do to know the real position of
the PossibleIdNLength data since we started reading. That allows a proper check
on the end value since that start.