delete half-constructed elements on read failure
When an EbmlMaster tries to read the next element, it uses
EbmlStream.FindNextElement() which only reads the EBML ID and the size
field. With that data it constructs the corresponding (Matroska)
element. It's now up to the EbmlMaster to read the element's payload.

If reading the payload fails, an exception will be thrown by
e.g. IOCallback.readFully(). So far the Master didn't catch that
exception, causing the half-constructed element to be dangling as it
hasn't been added to the Master yet at that point. Therefore the
exception must be caught, the element deleted and the exception
re-thrown.