EbmlCrc32: add explicit assignment operator
The class contains an explicit copy constructor. C++ used to
auto-generate a default assignment operator from that if none
existed. However, that behavior is now deprecated.

Another way could have been removing the existing copy constructor as
all it did was an element-wise copy of the source which is also the
default behavior. However, removing an existing constructor can break
the ABI while adding a non-virtual member function does not.