EbmlElement::FindNextElement: fix buffer overflow
If the first byte read has its upper four bits unset, the old code
would enter the inner `while` a fifth time as the old check might
increase `PossibleID_Length` to four and only breaking if the
resulting value was `> 4`. In that fifth iteration the next byte would
be read intothe fifth byte of the `PossibleId` array, a four-byte
array.