fix overflow reading on 32 bits system
The positions in libebml are stored in filepos_t which is in 64 bits. But when
reading in memory we are limited to 32 bits on 32 bits systems. We cannot
allocate more and read into such buffer.

When the size to read is bigger than what the system can handle, just assume
it's an allocation failure (it won't be because the 64 bits value will be
casted to 32 bits).

Ref. CVE-2021-3405