change return type of IOCallback::read() to size_t
Fixes #108.

The function is supposed to return the number of bytes actually
read. As it takes a "number of bytes to read" parameter of type
`size_t`, it cannot signal success properly if the number of bytes to
read exceeds the size of its former return type, `uint32`, which is 4
GB.

This breaks the API & the ABI. Users of this class hierarchy have to
adjust their overrides as well.