MemIOCallback: fix buffer overflow when reading too much data
If the addition of 2 positive values is smaller than one of the values then we
have an overflowing addition.

In this case that means we are trying to read more data that is actually in
our buffer. So we can use the same mechanism as reading too much data.

(cherry picked from commit 4d577f5c3e267b2988d56dafebc82dedb4c45506)
Signed-off-by: Steve Lhomme <slhomme@matroska.org>