use a templated type for default values
So the default value is hardcoded in the class semantic info and not
set in the constructor.

A new EbmlElement subclass is created for EbmlElement types that may
have a default value.

An IsSameValue() method is added to compare an element to a value. This
is used by the EbmlElementDefault class to tell if the element value is
the same as the default when there is one.