Don't assert when calling incorrectly SetSizeInfinite()
It is called in FindNextId/FindNextElement which always checks for failure.
So don't need the assert there.

It's also called from the EbmlMaster constructor.
The copy constructor is fine since the the original can only have a valid state.
The other constructor has a default value of finite size and the macros never set this to infinite.
In libmatroska the KaxSegment constructor calls SetSizeInfinite() rather than using the default value.
But we don't need an assert there, since we know it's allowed.