remove EbmlDummy default constructor
The whole point of EbmlDummy is to be a placeholder for elements with an unknown EBML ID.
It should not exist with a fake ID not corresponding to something read.
The current API makes some default creation mandatory but with don't have to allow
a local creation on top of that.