add a developer mode option
This triggers more compilation warning.

It is not on by default so the code will keep building without new warning in the future.
It is meant for developers of the library to produce cleaner code and/or detect issues early.

(cherry picked from commit fd4ad0cb08525c2a8a77f0f6391e33df0df51777)